requirements.txt
manim 
streamlit 
protobuf 
openai 
moviepy 
matplotlib 
google-cloud
pillow 
decorator 
google-cloud-texttospeech
imageio 
ffmpeg
imageio[ffmpeg]
imageio_ffmpeg
